feature/ARCWELL-412-fix-2 - Moved detail.* to feature level
