+++ packages/admin/src/app/feature/detail/detail.component.spec.ts
