+++ packages/admin/src/app/feature/detail/detail.store.ts
