+++ packages/admin/src/app/feature/person-type/person-type.component.ts
@@ -37,7 +37,7 @@ import { TagsFormComponent } from '@shared/components/tags-form/tags-form.compon
 import { autoSlugify } from '@shared/helpers/auto-slug.helper';
 import { BackButtonComponent } from '@shared/components/back-button/back-button.component';
 import { DetailHeaderComponent } from '@shared/components/detail-header/detail-header.component';
-import { DetailStore } from '@feature/project-management/detail/detail.store';
+import { DetailStore } from '@feature/detail/detail.store';
 
 @Component({
   selector: 'aw-person-type',
