+++ packages/admin/src/app/feature/resource/resource.component.ts
@@ -32,7 +32,7 @@ import { ResourceTypeType } from '@app/shared/schemas/resource-type.schema';
 import { ErrorContainerComponent } from '../error-container/error-container.component';
 import { BackButtonComponent } from '@app/shared/components/back-button/back-button.component';
 import { DetailHeaderComponent } from '@shared/components/detail-header/detail-header.component';
-import { DetailStore } from '@feature/project-management/detail/detail.store';
+import { DetailStore } from '@feature/detail/detail.store';
 import {
   ResourceNewType,
   ResourceType,
